Addition of optional parameters -c and -t
It was used the getopt library to implement the following parameters.
The parameter -c means that the program output will be redirected to a csv file with the appropriate format (ex: ./energy_storms -c out.csv 50 test_files/*).
The parameter -t specifies the number of omp threads assigned to the program.